[김수찬] - Exception, Swagger 문서 처리 중

+++ src/main/java/com/green/glampick/service/implement/AdminServiceImpl.java
@@ -2,6 +2,7 @@ package com.green.glampick.service.implement;
 
 import com.green.glampick.common.CustomFileUtils;
 import com.green.glampick.common.Role;
+import com.green.glampick.dto.request.admin.DeleteBannerRequestDto;
 import com.green.glampick.dto.request.admin.module.AdminModule;
 import com.green.glampick.dto.request.owner.module.RoomModule;
 import com.green.glampick.dto.response.admin.*;
@@ -17,6 +18,7 @@ import com.green.glampick.repository.*;
 import com.green.glampick.repository.resultset.GetAccessGlampingListResultSet;
 import com.green.glampick.repository.resultset.GetAccessOwnerSignUpListResultSet;
 import com.green.glampick.repository.resultset.GetDeleteOwnerListResultSet;
+import com.green.glampick.security.AuthenticationFacade;
 import com.green.glampick.service.AdminService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -33,6 +35,7 @@ import static com.green.glampick.common.GlobalConst.MAX_BANNER_SIZE;
 @Service
 @RequiredArgsConstructor
 public class AdminServiceImpl implements AdminService {
+    private final AuthenticationFacade authenticationFacade;
     private final OwnerRepository ownerRepository;
     private final BannerRepository bannerRepository;
     private final GlampingRepository glampingRepository;
@@ -161,12 +164,25 @@ public class AdminServiceImpl implements AdminService {
     //  관리자 페이지 - 메인 화면 배너 삭제하기  //
     @Override
     @Transactional
-    public ResponseEntity<? super DeleteBannerResponseDto> deleteBanner(Long bannerId) {
+    public ResponseEntity<? super DeleteBannerResponseDto> deleteBanner(DeleteBannerRequestDto dto) {
+
+        try {
+            dto.setAdminIdx(authenticationFacade.getLoginUserId());
+            if (dto.getAdminIdx() <= 0) {
+                throw new RuntimeException();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new CustomException(CommonErrorCode.MNF);
+        }
+
 
         try {
 
-            BannerEntity bannerEntity = bannerRepository.findByBannerId(bannerId);
-            AdminModule.deleteImageOne(bannerId, bannerRepository, customFileUtils);
+            if (dto.getBannerId() == null) { throw new CustomException(AdminErrorCode.NFB); }
+
+            BannerEntity bannerEntity = bannerRepository.findByBannerId(dto.getBannerId());
+            AdminModule.deleteImageOne(dto.getBannerId(), bannerRepository, customFileUtils);
             bannerRepository.delete(bannerEntity);
 
         } catch (CustomException e) {
@@ -177,7 +193,6 @@ public class AdminServiceImpl implements AdminService {
         }
 
         return DeleteBannerResponseDto.success();
-
 
     }
